# Sort todos before executing them