Improves fairness of STM by sorting todos before executing them. To determine how much performance dropped, I ran two "standard" benchmarking suites for STM - TMap operations and a deep flat map.

### Before the change

Deep flat map:

```
Benchmark                        (depth)   Mode  Cnt      Score     Error  Units
STMFlatMapBenchmark.deepFlatMap       20  thrpt   15    888.181 ±    4.118  ops/s
```

TMap operations:

```
Benchmark                     (size)   Mode  Cnt      Score     Error  Units
TMapOpsBenchmarks.lookup       10000  thrpt   15   3782.983 ±  22.903  ops/s
TMapOpsBenchmarks.removal      10000  thrpt   15   2651.815 ±  13.043  ops/s
TMapOpsBenchmarks.transform    10000  thrpt   15     27.559 ±   0.163  ops/s
TMapOpsBenchmarks.transformM   10000  thrpt   15    21.695 ±   0.146  ops/s
TMapOpsBenchmarks.update       10000  thrpt   15   2014.471 ±  14.065  ops/s
```

### After the change

Deep flat map:

```
Benchmark                        (depth)   Mode  Cnt      Score     Error  Units
STMFlatMapBenchmark.deepFlatMap       20  thrpt   15    870.717 ±   40.045  ops/s
```

TMap operations:

```
Benchmark                     (size)   Mode  Cnt      Score     Error  Units
TMapOpsBenchmarks.lookup       10000  thrpt   15   3914.755 ±  19.434  ops/s
TMapOpsBenchmarks.removal      10000  thrpt   15   2376.030 ±  19.728  ops/s
TMapOpsBenchmarks.transform    10000  thrpt   15     27.365 ±   0.216  ops/s
TMapOpsBenchmarks.transformM   10000  thrpt   15     25.187 ±   0.154  ops/s
TMapOpsBenchmarks.update       10000  thrpt   15   1945.318 ±  15.071  ops/s
```