@jdegoes 2020-04-11
@mijicd Do you think we could come up with a test of fairness?