@mijicd 2020-04-11
After rethinking this idea, I've decided it's better to close this PR. While our benchmarks show slight performance hit, insisting on fairness under contention would be way more expensive. To the best of my knowledge, the majority of Java's concurrency primitives, as well as Haskell's STM do not support fairness. 

One notable exception to this is Java's ReentrantLock that allows setting the fairness parameter, and it might make sense to expand `TReentrantLock` to support it.